reorganize chart modularity & run clippy
