+++ src/chart/indicator/open_interest.rs
@@ -4,13 +4,13 @@ use iced::widget::canvas::{self, Cache, Event, Geometry, Path, Stroke};
 use iced::widget::{Canvas, center, container, row, text, vertical_rule};
 use iced::{Element, Length, Point, Rectangle, Renderer, Size, Theme, Vector, mouse};
 
-use crate::chart::{Basis, Caches, CommonChartData, Interaction, Message};
+use crate::chart::{Basis, Caches, ViewState, Interaction, Message};
 use crate::style::{self, dashed_line};
 use data::util::{format_with_commas, guesstimate_ticks, round_to_tick};
 use exchange::Timeframe;
 
 pub fn indicator_elem<'a>(
-    chart_state: &'a CommonChartData,
+    chart_state: &'a ViewState,
     cache: &'a Caches,
     datapoints: &'a BTreeMap<u64, f32>,
     earliest: u64,
@@ -80,7 +80,7 @@ pub fn indicator_elem<'a>(
 pub struct OpenInterest<'a> {
     pub indicator_cache: &'a Cache,
     pub crosshair_cache: &'a Cache,
-    pub chart_state: &'a CommonChartData,
+    pub chart_state: &'a ViewState,
     pub max_value: f32,
     pub timeseries: &'a BTreeMap<u64, f32>,
 }
